refactor(game-engine): split actionHandler by concern
Extract from actionHandler.ts (657 -> 236 dispatch shell):
- revealPayload.ts: reveal handler registry + payload builder
- actionGuards.ts: precondition validation + nightmare block guard
- viewedRoleHandler.ts: VIEWED_ROLE intent handler

Re-exports from actionHandler.ts for backward compatibility.
